Moved NodeFactory stuff out of Registry
