+++ src/main/java/com/marginallyclever/ro3/Registry.java
@@ -5,31 +5,9 @@ import com.marginallyclever.ro3.mesh.MeshFactory;
 import com.marginallyclever.ro3.node.Node;
 import com.marginallyclever.ro3.node.nodefactory.NodeFactory;
 import com.marginallyclever.ro3.node.nodes.*;
-import com.marginallyclever.ro3.node.nodes.behavior.BehaviorTreeRunner;
-import com.marginallyclever.ro3.node.nodes.behavior.Fallback;
-import com.marginallyclever.ro3.node.nodes.behavior.Sequence;
-import com.marginallyclever.ro3.node.nodes.behavior.actions.LimbMoveToTarget;
 import com.marginallyclever.ro3.node.nodes.behavior.decorators.*;
-import com.marginallyclever.ro3.node.nodes.crab.Crab;
-import com.marginallyclever.ro3.node.nodes.environment.Environment;
-import com.marginallyclever.ro3.node.nodes.limbplanner.LimbPlanner;
-import com.marginallyclever.ro3.node.nodes.limbsolver.LimbSolver;
-import com.marginallyclever.ro3.node.nodes.marlinrobot.linearstewartplatform.LinearStewartPlatform;
-import com.marginallyclever.ro3.node.nodes.marlinrobot.marlinrobotarm.MarlinRobotArm;
-import com.marginallyclever.ro3.node.nodes.Material;
-import com.marginallyclever.ro3.node.nodes.networksession.NetworkSession;
-import com.marginallyclever.ro3.node.nodes.neuralnetwork.Brain;
-import com.marginallyclever.ro3.node.nodes.neuralnetwork.Neuron;
-import com.marginallyclever.ro3.node.nodes.neuralnetwork.Synapse;
-import com.marginallyclever.ro3.node.nodes.neuralnetwork.leglimbic.LegLimbic;
 import com.marginallyclever.ro3.node.nodes.odenode.*;
-import com.marginallyclever.ro3.node.nodes.odenode.odebody.odebodies.ODEBox;
-import com.marginallyclever.ro3.node.nodes.odenode.odebody.odebodies.ODECapsule;
-import com.marginallyclever.ro3.node.nodes.odenode.odebody.odebodies.ODECylinder;
-import com.marginallyclever.ro3.node.nodes.odenode.odebody.odebodies.ODESphere;
-import com.marginallyclever.ro3.node.nodes.pose.Pose;
 import com.marginallyclever.ro3.node.nodes.pose.poses.*;
-import com.marginallyclever.ro3.node.nodes.pose.poses.space.SpaceShip;
 import com.marginallyclever.ro3.physics.ODEPhysics;
 import com.marginallyclever.ro3.texture.TextureFactory;
 
@@ -54,71 +32,7 @@ public class Registry {
     public static void start() {
         nodeFactory.clear();
         nodeFactory.scan("com.marginallyclever.ro3");
-        NodeFactory.Category node = nodeFactory.getRoot();
-        {
-            NodeFactory.Category behavior = node.add("Behavior", null);
-            {
-                NodeFactory.Category action = behavior.add("Action", null);
-                {
-                    action.add("LimbMoveToTarget", LimbMoveToTarget::new);
-                }
-                behavior.add("Control", null);
-                NodeFactory.Category decorator = behavior.add("Decorator", null);
-                {
-                    decorator.add("ForceFailure", ForceFailure::new);
-                    decorator.add("ForceSuccess", ForceSuccess::new);
-                    decorator.add("Inverter", Inverter::new);
-                    decorator.add("KeepRunningUntilFailure", KeepRunningUntilFailure::new);
-                    decorator.add("Repeat", Repeat::new);
-                    decorator.add("RetryUntilSuccessful", RetryUntilSuccessful::new);
-                }
-                behavior.add("Fallback", Fallback::new);
-                behavior.add("Sequence", Sequence::new);
-            }
-            node.add("BehaviorTreeRunner", BehaviorTreeRunner::new);
-            node.add("Crab", Crab::new);
-            node.add("DHParameter", DHParameter::new);
-            node.add("Environment", Environment::new);
-            node.add("HingeJoint", HingeJoint::new);
-            node.add("LimbPlanner", LimbPlanner::new);
-            node.add("LimbSolver", LimbSolver::new);
-            node.add("LinearStewartPlatform", LinearStewartPlatform::new);
-            node.add("LinearJoint", LinearJoint::new);
-            node.add("MarlinRobotArm", MarlinRobotArm::new);
-            node.add("Material", Material::new);
-            node.add("Motor", Motor::new);
-            node.add("NetworkSession", NetworkSession::new);
-            NodeFactory.Category nn = node.add("NeuralNetwork", null);
-            {
-                nn.add("Brain", Brain::new);
-                nn.add("LegLimbic", LegLimbic::new);
-                nn.add("Neuron", Neuron::new);
-                nn.add("Synapse", Synapse::new);
-            }
-            NodeFactory.Category pose = node.add("Pose", Pose::new);
-            {
-                pose.add("AttachmentPoint", AttachmentPoint::new);
-                pose.add("Camera", Camera::new);
-                pose.add("Limb", Limb::new);
-                pose.add("LookAt", LookAt::new);
-                pose.add("MeshInstance", MeshInstance::new);
-                pose.add("SpaceShip", SpaceShip::new);
-            }
-            NodeFactory.Category physics = node.add("Physics", null);
-            {
-                physics.add("CreatureController", CreatureController::new);
-                physics.add("ODEAngularMotor", ODEAngularMotor::new);
-                physics.add("ODEBallSocket", ODEBallSocket::new);
-                physics.add("ODEBox", ODEBox::new);
-                physics.add("ODECapsule", ODECapsule::new);
-                physics.add("ODECylinder", ODECylinder::new);
-                physics.add("ODEHinge", ODEHinge::new);
-                physics.add("ODEPlane", ODEPlane::new);
-                physics.add("ODESlider", ODESlider::new);
-                physics.add("ODESphere", ODESphere::new);
-            }
-        }
-
+        nodeFactory.addKnownNodes();
         reset();
     }
