+++ src/main/java/com/marginallyclever/ro3/apps/nodetreeview/NodeTreeBranchRenderer.java
@@ -1,7 +1,5 @@
-package com.marginallyclever.ro3.node.nodefactory;
+package com.marginallyclever.ro3.apps.nodetreeview;
 
-import com.marginallyclever.ro3.apps.nodetreeview.NodeTreeBranch;
-import com.marginallyclever.ro3.apps.nodetreeview.NodeTreeView;
 import com.marginallyclever.ro3.node.Node;
 
 import javax.swing.*;
